backend/requirements.txt
django
djangorestframework
django-cors-headers
djangorestframework-simplejwt
mysqlclient
python-dotenv
django-filter
Pillow
drf-yasg
